Problemset 7 solved
Task 1
We implemented dependency injection mostly via
factories and in two cases with a provider. Because
many of our constructors needed parameters, this
aproach seemed more reasonable.

Task 2
Because we always require all constructor parameters
and we have no overloaded constructors, we chose not
to implement builders. The dependency injection requires
us to use many parameters, but we don't think builders
would be appropriate in this case.
We never implemented equals, so we din't implement hashCode.
We added a simple toString method for all game classes.